:seedling: add check-conditions to e2e tests.
