+++ test/e2e/caph.go
@@ -21,23 +21,14 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
-	"time"
 
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
-	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/utils/ptr"
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/cluster-api/test/framework"
 	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
 	"sigs.k8s.io/cluster-api/util"
-	"sigs.k8s.io/cluster-api/util/conditions"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	infrav1 "github.com/syself/cluster-api-provider-hetzner/api/v1beta1"
 )
 
 // CaphClusterDeploymentSpecInput is the input for CaphClusterDeploymentSpec.
@@ -50,86 +41,6 @@ type CaphClusterDeploymentSpecInput struct {
 	Flavor                   string
 	WorkerMachineCount       int64
 	ControlPlaneMachineCount int64
-}
-
-func logHCloudMachineStatusContinously(ctx context.Context, c client.Client) {
-	for {
-		t := time.After(30 * time.Second)
-		select {
-		case <-ctx.Done():
-			return
-		case <-t:
-			err := logHCloudMachineStatus(ctx, c)
-			if err != nil {
-				ginkgo.By(fmt.Sprintf("Error logging HCloudMachine status: %v", err))
-			}
-		}
-	}
-}
-
-func logHCloudMachineStatus(ctx context.Context, c client.Client) error {
-	hmList := &infrav1.HCloudMachineList{}
-	err := c.List(ctx, hmList)
-	if err != nil {
-		return err
-	}
-
-	if len(hmList.Items) == 0 {
-		return nil
-	}
-
-	caphDeployment := appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "caph-controller-manager",
-			Namespace: "caph-system",
-		},
-	}
-	err = c.Get(ctx, client.ObjectKeyFromObject(&caphDeployment), &caphDeployment)
-	if err != nil {
-		return fmt.Errorf("failed to get caph-controller-manager deployment: %w", err)
-	}
-
-	ginkgo.By(fmt.Sprintf("--------------------------------------------------- HCloudMachines %s",
-		caphDeployment.Spec.Template.Spec.Containers[0].Image))
-
-	for i := range hmList.Items {
-		hm := &hmList.Items[i]
-		if hm.Status.InstanceState == nil || *hm.Status.InstanceState == "" {
-			continue
-		}
-		addresses := make([]string, 0)
-		for _, a := range hm.Status.Addresses {
-			addresses = append(addresses, a.Address)
-		}
-
-		id := ""
-		if *hm.Spec.ProviderID != "" {
-			id = *hm.Spec.ProviderID
-		}
-		ginkgo.By("HCloudMachine: " + hm.Name + " " + id + " " + strings.Join(addresses, " "))
-		ginkgo.By("  ProvisioningState: " + string(*hm.Status.InstanceState))
-		l := make([]string, 0)
-		if hm.Status.FailureMessage != nil {
-			l = append(l, *hm.Status.FailureMessage)
-		}
-		if hm.Status.FailureMessage != nil {
-			l = append(l, *hm.Status.FailureMessage)
-		}
-		if len(l) > 0 {
-			ginkgo.By("  Error: " + strings.Join(l, ", "))
-		}
-		readyC := conditions.Get(hm, clusterv1.ReadyCondition)
-		msg := ""
-		reason := ""
-		state := "?"
-		if readyC != nil {
-			msg = readyC.Message
-			reason = readyC.Reason
-			state = string(readyC.Status)
-		}
-		ginkgo.By("  Ready Condition: " + state + " " + reason + " " + msg)
-	}
-	return nil
 }
 
 // CaphClusterDeploymentSpec implements a test that verifies that MachineDeployment rolling updates are successful.
@@ -162,7 +73,6 @@ func CaphClusterDeploymentSpec(ctx context.Context, inputGetter func() CaphClust
 
 	ginkgo.It("Should successfully create a cluster with three control planes", func() {
 		ginkgo.By("Creating a workload cluster")
-		go logHCloudMachineStatusContinously(ctx, input.BootstrapClusterProxy.GetClient())
 		clusterctl.ApplyClusterTemplateAndWait(ctx, clusterctl.ApplyClusterTemplateAndWaitInput{
 			ClusterProxy: input.BootstrapClusterProxy,
 			ConfigCluster: clusterctl.ConfigClusterInput{
